refactor(rest-contract): move DalIdCodec to the version-agnostic contract root
The codec has no coupling to v1 types and its encoding technique
(simple ids raw, composite ids as URL-safe unpadded Base64 of JSON)
is reusable by future contract versions. The /dal/v1 contract still
freezes its behavior while it binds to it, as documented in the
module's compatibility policy. Consumers only needed import updates.
