+++ telaio-rest-client/src/main/java/com/paganbit/telaio/rest/client/blocking/DalRestClient.java
@@ -11,8 +11,8 @@ import com.paganbit.telaio.rest.client.internal.DalErrorMapper;
 import com.paganbit.telaio.rest.client.internal.DalPayloadCodec;
 import com.paganbit.telaio.rest.client.internal.DalUriFactory;
 import com.paganbit.telaio.rest.client.internal.DalUriFactory.DalRequestUri;
+import com.paganbit.telaio.rest.contract.DalIdCodec;
 import com.paganbit.telaio.rest.contract.DalIdCodecException;
-import com.paganbit.telaio.rest.contract.v1.DalIdCodec;
 import com.turkraft.springfilter.converter.FilterStringConverter;
 import com.turkraft.springfilter.parser.node.FilterNode;
 import org.jspecify.annotations.Nullable;
